[Refactor] 저장 기능 detail 페이지에서만 구현될 수 있도록 함. #3
